Add New feature

+++ Parcel.Handling.Infra/Context.cs
@@ -15,26 +15,71 @@ namespace Parcel.Handling.Infra
 
         public Context(ApiContext context) =>
             (_context) = ( context);
-        
 
-        public async Task DeleteDepartmentById(int id)
+        public List<Department> GetDepartmentList()
+        {
+            var result = _context.Users.OrderBy( x=> x.Id).ToList();
+            return result;
+        }
+        public Task DeleteDepartmentById(int id)
         {
             var department = _context.Users.FirstOrDefault(i => i.Id == id);
-            _context.Users.Remove(department);
-            _context.SaveChanges();
+            if (department != null)
+                _context.Users.Remove(department);
+                _context.SaveChanges();
+            return Task.CompletedTask;
         }
-        public async Task SetNewDepartment(DepartmentDto department)
+        public Task AddNewDepartment(DepartmentDto department)
         {
             
             var addDepartment = new Department
             {
-                Id = department.Id,
                 Name = department.Name
             };
 
             _context.Users.Add(addDepartment);
             _context.SaveChanges();
+            return Task.CompletedTask;
+        }
 
+        public Task AddNewFistDepartment()
+        {
+
+
+            var department1 = new Department
+            {
+                Id = 1,
+                Name = "Mail"
+            };
+
+            _context.Users.Add(department1);
+
+            var department2 = new Department
+            {
+                Id = 2,
+                Name = "Regular"
+            };
+
+            _context.Users.Add(department2);
+
+            var department3 = new Department
+            {
+                Id = 3,
+                Name = "Heavy"
+            };
+
+            _context.Users.Add(department3);
+
+            var department4 = new Department
+            {
+                Id = 4,
+                Name = "Insurance"
+            };
+
+            _context.Users.Add(department4);
+            _context.SaveChanges();
+
+            return Task.CompletedTask;
         }
     }
 }
